Create lingumania-bsi.js
Created a somewhat strippped down, simplified version of Lingumania. Does not require (or use) the languages node in the linguJSON, but instead replaces that with a simple flag (translation_enabled) that allow you to enable/disable translation on any site that the file is included on. Removed the language switcher dropdown as well.