Create jQuery-Convert Secondary Nav Links to Modals.js
Converts all secondary nav item links that include "IsPopup=true" in the URL to pop up in a dialog.
    - Will not process nav items already marked to open in a new window (target=_blank)
    - Opens the modal window at 90% of the window height and width
    - Refreshes the main page on closing the modal window/popup
Place this code in a Content HTML iPart in a piece of content on the page. To apply to all pages, use footer content in the ContentAreas folder for the site.